+++ src/RefDocGen/CodeElements/Abstract/Types/IObjectTypeData.cs
@@ -1,6 +1,5 @@
 using RefDocGen.CodeElements.Abstract.Members;
 using RefDocGen.CodeElements.Abstract.Types.Enum;
-using System.Xml.Linq;
 
 namespace RefDocGen.CodeElements.Abstract.Types;
 
@@ -10,7 +9,7 @@ namespace RefDocGen.CodeElements.Abstract.Types;
 /// Note: This interface doesn't represent enum types - see <see cref="IEnumTypeData"/>.
 /// </para>
 /// </summary>
-public interface IObjectTypeData : ITypeNameData, ITypeDeclaration
+public interface IObjectTypeData : ITypeDeclaration
 {
     /// <summary>
     /// Checks if the type is abstract.
@@ -57,14 +56,4 @@ public interface IObjectTypeData : ITypeNameData, ITypeDeclaration
     /// Collection of indexers declared in the type.
     /// </summary>
     IReadOnlyList<IIndexerData> Indexers { get; }
-
-    /// <summary>
-    /// Base type of the type. Returns null, if the type has no base type (i.e. it's an interface or <see cref="object"/> type).
-    /// </summary>
-    ITypeNameData? BaseType { get; }
-
-    /// <summary>
-    /// Interfaces implemented by the current type.
-    /// </summary>
-    IReadOnlyList<ITypeNameData> Interfaces { get; }
 }
