+++ src/RefDocGen/CodeElements/Abstract/Types/ITypeDeclaration.cs
@@ -5,7 +5,7 @@ namespace RefDocGen.CodeElements.Abstract.Types;
 /// <summary>
 /// Represents a type declared in any of the assemblies analyzed.
 /// </summary>
-public interface ITypeDeclaration : ITypeNameBaseData
+public interface ITypeDeclaration : ITypeNameData
 {
     /// <summary>
     /// Access modifier of the type.
@@ -18,9 +18,14 @@ public interface ITypeDeclaration : ITypeNameBaseData
     XElement DocComment { get; }
 
     /// <summary>
-    /// Checks whether the type has any type parameters.
+    /// Base type of the type. Returns null, if the type has no base type (i.e. it's an interface or <see cref="object"/> type).
     /// </summary>
-    bool HasTypeParameters { get; }
+    ITypeNameData? BaseType { get; }
+
+    /// <summary>
+    /// Interfaces implemented by the current type.
+    /// </summary>
+    IReadOnlyList<ITypeNameData> Interfaces { get; }
 
     /// <summary>
     /// Collection of generic type parameters declared in the delegate, ordered by their index.
