+++ src/RefDocGen/CodeElements/Concrete/Types/Delegate/DelegateTypeData.cs
@@ -5,14 +5,13 @@ using RefDocGen.Tools.Xml;
 using System.Reflection;
 using System.Xml.Linq;
 using RefDocGen.CodeElements.Concrete.Members;
-using RefDocGen.CodeElements.Tools;
 
 namespace RefDocGen.CodeElements.Concrete.Types.Delegate;
 
 /// <summary>
 /// Class representing data of a delegate.
 /// </summary>
-internal class DelegateTypeData : TypeNameData, IDelegateTypeData
+internal class DelegateTypeData : TypeDeclaration, IDelegateTypeData
 {
     /// <summary>
     /// The method used for delegate invocation (i.e. <c>Invoke</c>).
@@ -28,36 +27,12 @@ internal class DelegateTypeData : TypeNameData, IDelegateTypeData
     public DelegateTypeData(Type type, MethodInfo invokeMethod, IReadOnlyDictionary<string, TypeParameterDeclaration> typeParameterDeclarations)
         : base(type, typeParameterDeclarations)
     {
-        TypeParameterDeclarations = typeParameterDeclarations;
 
         this.invokeMethod = new MethodData(invokeMethod, typeParameterDeclarations);
     }
 
     /// <inheritdoc/>
-    public override string Id => TypeId.Of(this);
-
-    /// <inheritdoc/>
-    public XElement DocComment { get; internal set; } = XmlDocElements.EmptySummary;
-
-    /// <summary>
-    /// Access modifier of the enum.
-    /// </summary>
-    public AccessModifier AccessModifier =>
-        AccessModifierExtensions.GetAccessModifier(
-            TypeObject.IsNestedPrivate,
-            TypeObject.IsNestedFamily,
-            TypeObject.IsNestedAssembly || TypeObject.IsNotPublic,
-            TypeObject.IsPublic || TypeObject.IsNestedPublic,
-            TypeObject.IsNestedFamANDAssem,
-            TypeObject.IsNestedFamORAssem);
-
-    /// <inheritdoc/>
     public XElement ReturnValueDocComment { get; internal set; } = XmlDocElements.EmptySummary;
-
-    /// <summary>
-    /// Collection of type parameters declared in the delegate; the keys represent type parameter names.
-    /// </summary>
-    public IReadOnlyDictionary<string, TypeParameterDeclaration> TypeParameterDeclarations { get; }
 
     /// <inheritdoc/>
     public ITypeNameData ReturnType => invokeMethod.ReturnType;
