+++ src/RefDocGen/CodeElements/Concrete/Types/TypeName/GenericParameterNameData.cs
@@ -1,7 +1,7 @@
 using RefDocGen.CodeElements.Abstract.Types;
 using RefDocGen.Tools;
 
-namespace RefDocGen.CodeElements.Concrete.Types;
+namespace RefDocGen.CodeElements.Concrete.Types.TypeName;
 
 /// <summary>
 /// Class representing name-related data of a generic parameter.
