+++ tests/RefDocGen.UnitTests/CodeElements/Concrete/Types/TypeNameDataTests.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using NSubstitute;
-using RefDocGen.CodeElements.Concrete.Types;
+using RefDocGen.CodeElements.Concrete.Types.TypeName;
 
 namespace RefDocGen.UnitTests.CodeElements.Concrete.Types;
